v1/backend/requirements.txt
SQLAlchemy
PyMySQL
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
pydantic
python-multipart
bcrypt